rewrite library to support python 3
