+++ .gitignore
@@ -58,3 +58,5 @@ target/
 
 # vim swp files
 *.swp
+
+.DS_Store
